Don't require PDF info at construction time, in preparation for making this settable by document.
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@5704 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/boxes-and-glue.cc
@@ -117,14 +117,14 @@ std::vector<std::vector<BoxesAndGlue::BoxOut>> BoxesAndGlue::PackBoxes(
       "topologically sorted, for one thing to inhibit cycles.";
     CHECK(box.parent_idx == -1 ||
           (box.parent_idx >= 0 &&
-           box.parent_idx < boxes.size()));
+           box.parent_idx < (int)boxes.size()));
     if (box.parent_idx == -1) {
       starting_nodes.push_back(i);
       depth.push_back(0);
     } else {
       CHECK(box.parent_idx >= 0) << box.parent_idx;
       successors[box.parent_idx].emplace_back(i, box.edge_penalty);
-      CHECK(box.parent_idx < depth.size());
+      CHECK(box.parent_idx < (int)depth.size());
       depth.push_back(depth[box.parent_idx] + 1);
     }
   }
